split tests in topic based files
