+++ definitions/npm/react-redux_v5.x.x/flow_v0.89.x-/test_connectExported.js
@@ -678,877 +678,3 @@ function checkIfStateTypeIsRespectedAgain() {
   <Connected />;
   e.push(Connected);
 }
-
-function testState() {
-  type Props = {};
-  class Com extends React.Component<Props> {}
-
-  type State = {|
-    a: number
-  |};
-  const mapStateToProps = (state: State) => {
-    return {}
-  };
-  type OtherState = {|
-    b: number
-  |};
-
-  //$ExpectError property `b` is missing in `State` but exists in `OtherState`
-  const Connected = connect<Props, {||}, _,_,OtherState,empty>(mapStateToProps)(Com);
-  e.push(Connected);
-  <Connected />;
-}
-
-function testDispatch() {
-  type Action1 = {|
-    type: string
-  |};
-  type Dispatch1 = Action1 => Action1;
-  const action1 = (): Action1 => ({ type: 'str' })
-  const mapDispatchToProps = {
-    action1
-  };
-
-  type Action2 = {|
-  //$ExpectError `string` is incompatible with `number` in property `type`
-    type: number
-  |};
-  type Dispatch2 = Action2 => Action2;
-
-  type Props = {
-    action1: typeof action1,
-  };
-  class Com extends React.Component<Props> {}
-
-  const Connected1 = connect<Props, {||}, _,_,_,Dispatch1>(null, mapDispatchToProps)(Com);
-  e.push(Connected1);
-  <Connected1 />;
-
-  const Connected2 = connect<Props, {||}, _,_,_,Dispatch2>(null, mapDispatchToProps)(Com);
-  e.push(Connected2);
-  <Connected2 />;
-}
-
-function testDispatchFn() {
-  type Action1 = {|
-    type: string
-  |};
-  type Dispatch1 = Action1 => Action1;
-  const action1 = (): Action1 => ({ type: 'str' });
-  type DispatchProps = {|
-    action1: () => Action1
-  |};
-  type MapDispatchToPropsFn = Dispatch1 => DispatchProps;
-  const mapDispatchToProps: MapDispatchToPropsFn = (dispatch: Dispatch1) => ({
-    action1: (...args) => dispatch(action1(...args))
-  });
-
-  type Action2 = {|
-  //$ExpectError `string` is incompatible with `number` in property `type`
-    type: number
-  |};
-  type Dispatch2 = Action2 => Action2;
-
-  type Props = {
-    action1: typeof action1,
-  };
-  class Com extends React.Component<Props> {}
-
-  const Connected1 = connect<Props, {||}, _,DispatchProps,_,Dispatch1>(null, mapDispatchToProps)(Com);
-  e.push(Connected1);
-  <Connected1 />;
-
-  const Connected2 = connect<Props, {||}, _,DispatchProps,_,Dispatch2>(null, mapDispatchToProps)(Com);
-  e.push(Connected2);
-  <Connected2 />;
-}
-
-function testDispatchFnThunk() {
-  type State = {||}
-  type Action = {|
-    type: 'action'
-  |};
-  type DispatchAction = Action => Action;
-  type Thunk = (Dispatch, State) => Promise<number>
-  type DispatchThunk = Thunk => Promise<number>
-  type Dispatch = DispatchAction & DispatchThunk
-
-  const action = (): Action => ({ type: 'action' });
-  const thunk = (): Thunk => (d: Dispatch) => Promise.resolve(1);
-  type DispatchProps = {|
-    action: typeof action,
-    thunk: () => Promise<number>,
-  |};
-  type MapDispatchToPropsFn = Dispatch => DispatchProps;
-  const mapDispatchToProps: MapDispatchToPropsFn = (dispatch) => ({
-    action: () => dispatch(action()),
-    thunk: () => dispatch(thunk())
-  });
-
-  type Props = {
-    // in the case of a function passed to `connect()` Flow infers
-    // the same type of function is both the Props and DispatchProps objects
-    // (see the testDispatchThunk() to get the difference)
-    ...DispatchProps
-  };
-  class Com extends React.Component<Props> {}
-
-  const Connected = connect<Props, {||}, _,DispatchProps,_,Dispatch>(null, mapDispatchToProps)(Com);
-  e.push(Connected);
-  <Connected />;
-}
-
-function testDispatchThunk_onlyDispatchObject_differentDispatchPropsAreOK() {
-  type State = {||}
-  type Action = {|
-    type: 'action'
-  |};
-  type DispatchAction = Action => Action;
-  type Thunk = (Dispatch, State) => Promise<number>
-  type DispatchThunk = Thunk => Promise<number>
-  type Dispatch = DispatchAction & DispatchThunk
-
-  const action = (): Action => ({ type: 'action' });
-  const thunk = (): Thunk => (d: Dispatch) => Promise.resolve(1);
-
-  type DispatchProps = {|
-    action: typeof action,
-    // here the property returns a thunk...
-    thunk: () => Thunk,
-  |};
-  const mapDispatchToProps = {
-    action,
-    thunk
-  };
-
-  type Props = {
-    action: typeof action,
-    // ... and here the property returns a return value of thunk
-    // as dispatch calls it for us with `dispatch` and `getState`
-    thunk: () => Promise<number>,
-  };
-  class Com extends React.Component<Props> {}
-
-  const Connected = connect<Props, {||}, _,DispatchProps,_,Dispatch>(null, mapDispatchToProps)(Com);
-  e.push(Connected);
-  <Connected />;
-}
-
-function testDispatchThunk_onlyDispatchObject_sameDispatchPropsAreErroneous() {
-  type State = {||}
-  type Action = {|
-    type: 'action'
-  |};
-  type DispatchAction = Action => Action;
-  type Thunk = (Dispatch, State) => Promise<number>
-  type DispatchThunk = Thunk => Promise<number>
-  type Dispatch = DispatchAction & DispatchThunk
-
-  const action = (): Action => ({ type: 'action' });
-  const thunk = (): Thunk => (d: Dispatch) => Promise.resolve(1);
-
-  type DispatchProps = {|
-    action: typeof action,
-    //$ExpectError here the property returns a thunk...
-    thunk: () => Thunk,
-  |};
-  const mapDispatchToProps = {
-    action,
-    thunk
-  };
-
-  type Props = {
-    // trying to pass the not passed to dispatch types (against the redux dispatch monad)
-    ...DispatchProps,
-  };
-  class Com extends React.Component<Props> {}
-
-  const Connected = connect<Props, {||}, _,DispatchProps,_,Dispatch>(null, mapDispatchToProps)(Com);
-  e.push(Connected);
-  <Connected />;
-}
-
-function testDispatchThunk_StateAndDispatchObject_differentDispatchPropsAreOK() {
-  type State = {|
-    state1: 'state1'
-  |}
-  type Action = {|
-    type: 'action'
-  |};
-  type DispatchAction = Action => Action;
-  type Thunk = (Dispatch, State) => Promise<number>
-  type DispatchThunk = Thunk => Promise<number>
-  type Dispatch = DispatchAction & DispatchThunk
-
-  const action = (): Action => ({ type: 'action' });
-  const thunk = (): Thunk => (d: Dispatch) => Promise.resolve(1);
-
-  type StateProps = {|
-    state1: 'state1',
-  |};
-  const mapStateToProps = state => ({
-    state1: state.state1
-  })
-
-  type DispatchProps = {|
-    action: typeof action,
-    // here the property returns a thunk...
-    thunk: () => Thunk,
-  |};
-  const mapDispatchToProps = {
-    action,
-    thunk
-  };
-
-  type Props = {
-    ...StateProps,
-    action: typeof action,
-    // ... and here the property returns a return value of thunk
-    // as dispatch calls it for us with `dispatch` and `getState`
-    thunk: () => Promise<number>,
-  };
-  class Com extends React.Component<Props> {}
-
-  const Connected = connect<Props, {||}, _,DispatchProps,_,Dispatch>(mapStateToProps, mapDispatchToProps)(Com);
-  e.push(Connected);
-  <Connected />;
-}
-
-function testDispatchThunk_StateAndDispatchObject_sameDispatchPropsAreErroneous() {
-  type State = {||}
-  type Action = {|
-    type: 'action'
-  |};
-  type DispatchAction = Action => Action;
-  type Thunk = (Dispatch, State) => Promise<number>
-  type DispatchThunk = Thunk => Promise<number>
-  type Dispatch = DispatchAction & DispatchThunk
-
-  const action = (): Action => ({ type: 'action' });
-  const thunk = (): Thunk => (d: Dispatch) => Promise.resolve(1);
-
-  type StateProps = {|
-    state1: 'state1',
-  |};
-  const mapStateToProps = state => ({
-    state1: state.state1
-  })
-
-  type DispatchProps = {|
-    action: typeof action,
-    //$ExpectError here the property returns a thunk...
-    thunk: () => Thunk,
-  |};
-  const mapDispatchToProps = {
-    action,
-    thunk
-  };
-
-  type Props = {
-    ...StateProps,
-    // trying to pass the not passed to dispatch types (against the redux dispatch monad)
-    ...DispatchProps,
-  };
-  class Com extends React.Component<Props> {}
-
-  const Connected = connect<Props, {||}, _,DispatchProps,_,Dispatch>(mapStateToProps, mapDispatchToProps)(Com);
-  e.push(Connected);
-  <Connected />;
-}
-
-function testMergeProps_OnlyOwnProps_ok() {
-  opaque type Action = 'action';
-  type Dispatch = Action => Action;
-
-  type OwnProps = {|
-    passthrough: string
-  |}
-  type Props = {
-    ...OwnProps
-  };
-  class Com extends React.Component<Props> {}
-
-  const mergeProps = (stateProps: {||}, dispatchProps: {|dispatch: Dispatch|}, ownProps: OwnProps) => {
-    return {
-      ...ownProps
-    }
-  }
-
-  const Connected = connect<Props, OwnProps, _,_,_,Dispatch>(null, null, mergeProps)(Com);
-  e.push(Connected);
-  <Connected passthrough="foo" />;
-}
-
-function testMergeProps_OnlyOwnProps_wrongDispatch() {
-  opaque type Action = 'action';
-  type Dispatch = Action => Action;
-
-  type OwnProps = {|
-    passthrough: string
-  |}
-  type Props = {
-    ...OwnProps
-  };
-  class Com extends React.Component<Props> {}
-
-  const mergeProps = (stateProps: {||}, dispatchProps: {|dispatch: string|}, ownProps: OwnProps) => {
-    return {
-      ...ownProps
-    }
-  }
-
-  const Connected = connect<Props, OwnProps, _,_,_,Dispatch>(
-    null,
-    null,
-    //$ExpectError string [1] is incompatible with  `Dispatch` [2] in property `dispatch`
-    mergeProps
-  )(Com);
-  e.push(Connected);
-  <Connected passthrough="foo" />;
-}
-
-function testMergeProps_OnlyOwnProps_noPassthrough() {
-  opaque type Action = 'action';
-  type Dispatch = Action => Action;
-
-  type OwnProps = {|
-    passthrough: string
-  |}
-  type Props = {
-    ...OwnProps
-  };
-  class Com extends React.Component<Props> {}
-
-  const mergeProps = (stateProps: {||}, dispatchProps: {|dispatch: Dispatch|}, ownProps: OwnProps) => {
-    return {
-      a: 1
-    }
-  }
-
-  const Connected = connect<Props, OwnProps, _,_,_,Dispatch>(
-    null,
-    null,
-    //$ExpectError property `passthrough` is missing in object literal [1] but exists in `OwnProps` [2]
-    mergeProps
-  )(Com);
-  e.push(Connected);
-  <Connected passthrough="foo" />;
-}
-
-function testMergeProps_OnlyStateProps_ok() {
-  type State = {|
-    +state1: 'state1'
-  |}
-  opaque type Action = 'action';
-  type Dispatch = Action => Action;
-
-  type OwnProps = {|
-    passthrough: string
-  |}
-  type StateProps = {|
-    state1: 'state1'
-  |}
-  type Props = {
-    ...OwnProps,
-    ...StateProps,
-  };
-  class Com extends React.Component<Props> {}
-
-  const mapStateToProps = state => ({state1: state.state1})
-
-  const mergeProps = (stateProps: StateProps, dispatchProps: {|dispatch: Dispatch|}, ownProps: OwnProps) => {
-    return {
-      ...ownProps,
-      ...stateProps,
-    }
-  }
-
-  const Connected = connect<Props, OwnProps, _,_,_,Dispatch>(mapStateToProps, null, mergeProps)(Com);
-  e.push(Connected);
-  <Connected passthrough="foo" />;
-}
-
-function testMergeProps_OnlyStateProps_wrongDispatch() {
-  type State = {|
-    +state1: 'state1'
-  |}
-  opaque type Action = 'action';
-  type Dispatch = Action => Action;
-
-  type OwnProps = {|
-    passthrough: string
-  |}
-  type StateProps = {|
-    state1: 'state1'
-  |}
-  type Props = {
-    ...OwnProps,
-    ...StateProps
-  };
-  class Com extends React.Component<Props> {}
-
-  const mapStateToProps = state => ({
-    state1: state.state1
-  })
-
-  const mergeProps = (stateProps: StateProps, dispatchProps: {|dispatch: string|}, ownProps: OwnProps) => {
-    return {
-      ...ownProps,
-      ...stateProps
-    }
-  }
-
-  const Connected = connect<Props, OwnProps, _,_,_,Dispatch>(
-    mapStateToProps,
-    null,
-    //$ExpectError string [1] is incompatible with  `Dispatch` [2] in property `dispatch`
-    mergeProps
-  )(Com);
-  e.push(Connected);
-  <Connected passthrough="foo" />;
-}
-
-function testMergeProps_OnlyDispatchPropsObject_ok() {
-  type State = {|
-    +state1: 'state1'
-  |}
-  opaque type Action = 'action1';
-  type Dispatch = Action => Action;
-  const action1 = (): Action => 'action1'
-
-  type OwnProps = {|
-    passthrough: string
-  |}
-  type DispatchProps = {|
-    action1: typeof action1
-  |}
-  type Props = {
-    ...OwnProps,
-    ...DispatchProps,
-  };
-  class Com extends React.Component<Props> {}
-
-  const mapDispatchToProps = {
-    action1
-  }
-
-  const mergeProps = (stateProps: {||}, dispatchProps: DispatchProps, ownProps: OwnProps) => {
-    return {
-      ...ownProps,
-      ...dispatchProps,
-    }
-  }
-
-  const Connected = connect<Props, OwnProps, _,_,_,Dispatch>(undefined, mapDispatchToProps, mergeProps)(Com);
-  e.push(Connected);
-  <Connected passthrough="foo" />;
-}
-
-function testMergeProps_OnlyDispatchPropsObject_wrongExpectedState() {
-  type State = {|
-    +state1: 'state1'
-  |}
-  opaque type Action = 'action1';
-  type Dispatch = Action => Action;
-  const action1 = (): Action => 'action1'
-
-  type OwnProps = {|
-    passthrough: string
-  |}
-  type DispatchProps = {|
-    action1: typeof action1
-  |}
-  type Props = {
-    ...OwnProps,
-    ...DispatchProps,
-  };
-  class Com extends React.Component<Props> {}
-
-  const mapDispatchToProps = {
-    action1
-  }
-
-  const mergeProps = (stateProps: {wrong: boolean}, dispatchProps: DispatchProps, ownProps: OwnProps) => {
-    return {
-      ...ownProps,
-      ...dispatchProps,
-    }
-  }
-
-  const Connected = connect<Props, OwnProps, _,_,_,Dispatch>(
-    undefined,
-    mapDispatchToProps,
-    //$ExpectError property `wrong` is missing in object type [1] but exists in object type [2]
-    mergeProps
-  )(Com);
-  e.push(Connected);
-  <Connected passthrough="foo" />;
-}
-
-function testMergeProps_OnlyDispatchPropsFunction_ok() {
-  type State = {|
-    +state1: 'state1'
-  |}
-  opaque type Action = 'action1';
-  type Dispatch = Action => Action;
-  const action1 = (): Action => 'action1'
-
-  type OwnProps = {|
-    passthrough: string
-  |}
-  type DispatchProps = {|
-    action1: typeof action1
-  |}
-  type Props = {
-    ...OwnProps,
-    ...DispatchProps,
-  };
-  class Com extends React.Component<Props> {}
-
-  const mapDispatchToPropsFn = dispatch => ({
-    action1: () => dispatch(action1())
-  })
-
-  const mergeProps = (stateProps: {||}, dispatchProps: DispatchProps, ownProps: OwnProps) => {
-    return {
-      ...ownProps,
-      ...dispatchProps,
-    }
-  }
-
-  const Connected = connect<Props, OwnProps, _,_,_,Dispatch>(null, mapDispatchToPropsFn, mergeProps)(Com);
-  e.push(Connected);
-  <Connected passthrough="foo" />;
-}
-
-function testMergeProps_OnlyDispatchPropsFunction_WrongExpectedState() {
-  type State = {|
-    +state1: 'state1'
-  |}
-  opaque type Action = 'action1';
-  type Dispatch = Action => Action;
-  const action1 = (): Action => 'action1'
-
-  type OwnProps = {|
-    passthrough: string
-  |}
-  type DispatchProps = {|
-    action1: typeof action1
-  |}
-  type Props = {
-    ...OwnProps,
-    ...DispatchProps,
-  };
-  class Com extends React.Component<Props> {}
-
-  type MapDispatchToPropsFn = Dispatch => DispatchProps
-  const mapDispatchToPropsFn: MapDispatchToPropsFn = dispatch => ({
-    action1: () => dispatch(action1())
-  })
-
-  const mergeProps = (stateProps: {|wrong:boolean|}, dispatchProps: DispatchProps, ownProps: OwnProps) => {
-    return {
-      ...ownProps,
-      ...dispatchProps,
-    }
-  }
-
-  //$ExpectError property `wrong` is missing in  object type [1] but exists in  object type [2]
-  const Connected = connect<Props, OwnProps, _,_,_,Dispatch>(null, mapDispatchToPropsFn, mergeProps)(Com);
-  e.push(Connected);
-  <Connected passthrough="foo" />;
-}
-
-function testMergeProps_OnlyDispatchPropsFunction_wrongDispatchProp() {
-  type State = {|
-    +state1: 'state1'
-  |}
-  opaque type Action = 'action1';
-  type Dispatch = Action => Action;
-  const action1 = (): Action => 'action1'
-
-  type OwnProps = {|
-    passthrough: string
-  |}
-  type DispatchProps = {|
-    action1: typeof action1
-  |}
-  type Props = {
-    ...OwnProps,
-    ...DispatchProps,
-  };
-  class Com extends React.Component<Props> {}
-
-  const mapDispatchToPropsFn = dispatch => ({
-    action1: () => 123
-  })
-
-  const mergeProps = (stateProps: {||}, dispatchProps: DispatchProps, ownProps: OwnProps) => {
-    return {
-      ...ownProps,
-      ...dispatchProps,
-    }
-  }
-
-  const Connected = connect<Props, OwnProps, _,_,_,Dispatch>(
-    null,
-    //$ExpectError number [1] is incompatible with  string literal `action1` [2]
-    mapDispatchToPropsFn,
-    mergeProps
-  )(Com);
-  e.push(Connected);
-  <Connected passthrough="foo" />;
-}
-
-function testMergeProps_StateAndDispatchPropsFunction_ok() {
-  type State = {|
-    +state1: 'state1'
-  |}
-  opaque type Action = 'action1';
-  type Dispatch = Action => Action;
-  const action1 = (): Action => 'action1'
-
-  type OwnProps = {|
-    passthrough: string
-  |}
-  type StateProps = {|
-    state1: 'state1'
-  |}
-  type DispatchProps = {|
-    action1: typeof action1
-  |}
-  type Props = {
-    ...OwnProps,
-    ...StateProps,
-    ...DispatchProps,
-  };
-  class Com extends React.Component<Props> {}
-
-  const mapStateToProps = state => ({
-    state1: state.state1
-  })
-
-  const mapDispatchToPropsFn = dispatch => ({
-    action1: () => dispatch(action1())
-  })
-
-  const mergeProps = (stateProps: StateProps, dispatchProps: DispatchProps, ownProps: OwnProps) => {
-    return {
-      ...ownProps,
-      ...stateProps,
-      ...dispatchProps,
-    }
-  }
-
-  const Connected = connect<Props, OwnProps, _,_,_,Dispatch>(
-    mapStateToProps,
-    mapDispatchToPropsFn,
-    mergeProps
-  )(Com);
-  e.push(Connected);
-  <Connected passthrough="foo" />;
-}
-
-function testMergeProps_StateAndDispatchPropsFunction_wrongState() {
-  type State = {|
-    +state1: 'state1'
-  |}
-  opaque type Action = 'action1';
-  type Dispatch = Action => Action;
-  const action1 = (): Action => 'action1'
-
-  type OwnProps = {|
-    passthrough: string
-  |}
-  type StateProps = {|
-    state1: 'state1'
-  |}
-  type DispatchProps = {|
-    action1: typeof action1
-  |}
-  type Props = {
-    ...OwnProps,
-    ...StateProps,
-    ...DispatchProps,
-  };
-  class Com extends React.Component<Props> {}
-
-  const mapStateToProps = state => ({
-    state1: state.state1
-  })
-
-  const mapDispatchToPropsFn = dispatch => ({
-    action1: () => dispatch(action1())
-  })
-
-  const mergeProps = (
-    // yes, a bit cryptic
-    //$ExpectError property `state1` is missing in  object type [1] but exists in  object literal [2]
-    stateProps: {|wrong:boolean|},
-    dispatchProps: DispatchProps,
-    ownProps: OwnProps
-  ) => {
-    return {
-      ...ownProps,
-      state1: 'state1',
-      ...dispatchProps,
-    }
-  }
-
-  const Connected = connect<Props, OwnProps, _,_,_,Dispatch>(
-    mapStateToProps,
-    mapDispatchToPropsFn,
-    mergeProps
-  )(Com);
-  e.push(Connected);
-  <Connected passthrough="foo" />;
-}
-
-function testMergeProps_StateAndDispatchPropsFunction_wrongDispatch() {
-  type State = {|
-    +state1: 'state1'
-  |}
-  opaque type Action = 'action1';
-  type Dispatch = Action => Action;
-  const action1 = (): Action => 'action1'
-
-  type OwnProps = {|
-    passthrough: string
-  |}
-  type StateProps = {|
-    state1: 'state1'
-  |}
-  type DispatchProps = {|
-    action1: typeof action1
-  |}
-  type Props = {
-    ...OwnProps,
-    ...StateProps,
-    ...DispatchProps,
-  };
-  class Com extends React.Component<Props> {}
-
-  const mapStateToProps = state => ({
-    state1: state.state1
-  })
-
-  const mapDispatchToPropsFn = dispatch => ({
-    //$ExpectError number [1] is incompatible with string literal `action1` [2]
-    action1: () => 123
-  })
-
-  const mergeProps = (stateProps: StateProps, dispatchProps: DispatchProps, ownProps: OwnProps) => {
-    return {
-      ...ownProps,
-      ...stateProps,
-      ...dispatchProps,
-    }
-  }
-
-  const Connected = connect<Props, OwnProps, _,_,_,Dispatch>(
-    mapStateToProps,
-    mapDispatchToPropsFn,
-    mergeProps
-  )(Com);
-  e.push(Connected);
-  <Connected passthrough="foo" />;
-}
-
-function testMergeProps_returnsTotallyDifferentProps() {
-  type State = {|
-    +state1: 'state1'
-  |}
-  opaque type Action = 'action1';
-  type Dispatch = Action => Action;
-  const action1 = (): Action => 'action1'
-
-  type OwnProps = {|
-    passthrough: string
-  |}
-  type StateProps = {|
-    state1: 'state1'
-  |}
-  type DispatchProps = {|
-    action1: typeof action1
-  |}
-  type Props = {
-    a: 1,
-    b: 2,
-    c: 3
-  };
-  class Com extends React.Component<Props> {}
-
-  const mapStateToProps = state => ({
-    state1: state.state1
-  })
-
-  const mapDispatchToPropsFn = dispatch => ({
-    action1: () => dispatch(action1())
-  })
-
-  const mergeProps = (stateProps: StateProps, dispatchProps: DispatchProps, ownProps: OwnProps) => {
-    return {
-      a: 1,
-      b: 2,
-      c: 3,
-      d: 4
-    }
-  }
-
-  const Connected = connect<Props, OwnProps, _,_,_,Dispatch>(
-    mapStateToProps,
-    mapDispatchToPropsFn,
-    mergeProps
-  )(Com);
-  e.push(Connected);
-  <Connected passthrough="foo" />;
-}
-
-function testMergeProps_returnsTotallyDifferentPropsWithError() {
-  type State = {|
-    +state1: 'state1'
-  |}
-  opaque type Action = 'action1';
-  type Dispatch = Action => Action;
-  const action1 = (): Action => 'action1'
-
-  type OwnProps = {|
-    passthrough: string
-  |}
-  type StateProps = {|
-    state1: 'state1'
-  |}
-  type DispatchProps = {|
-    action1: typeof action1
-  |}
-  type Props = {
-    a: 1,
-    b: 2,
-    c: 3
-  };
-  class Com extends React.Component<Props> {}
-
-  const mapStateToProps = state => ({
-    state1: state.state1
-  })
-
-  const mapDispatchToPropsFn = dispatch => ({
-    action1: () => dispatch(action1())
-  })
-
-  const mergeProps = (stateProps: StateProps, dispatchProps: DispatchProps, ownProps: OwnProps) => {
-    //$ExpectError property `c` is missing in object literal [1] but exists in  `Props` [2]
-    return {
-      a: 1,
-      b: 2,
-    }
-  }
-
-  const Connected = connect<Props, OwnProps, _,_,_,Dispatch>(
-    mapStateToProps,
-    mapDispatchToPropsFn,
-    mergeProps
-  )(Com);
-  e.push(Connected);
-  <Connected passthrough="foo" />;
-}
